+++ abacusflow-usecase/abacusflow-usecase-inventory/src/main/kotlin/org/bruwave/abacusflow/usecase/inventory/BasicInventoryUnitTO.kt
@@ -1,5 +1,6 @@
 package org.bruwave.abacusflow.usecase.inventory
 
+import java.math.BigDecimal
 import java.time.Instant
 import java.util.UUID
 
@@ -12,7 +13,7 @@ data class BasicInventoryUnitTO(
     val depotName: String?,
     val quantity: Long,
     val remainingQuantity: Long,
-    val unitPrice: Double,
+    val unitPrice: BigDecimal,
     val receivedAt: Instant,
     val batchCode: UUID?, // 如果是批次单元
     val serialNumber: String?, // 如果是资产单元
